separated state logic from main
